// examples/client/src/gen7/models/ts/petController/AddPet.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { AddPetRequest } from '../AddPetRequest.ts'
import type { Pet } from '../Pet.ts'

/**
 * @type object
 */
export type AddPetStatus200Json = Pet

/**
 * @type object
 */
export type AddPetStatus200Xml = Pet

export type AddPetStatus200 = AddPetStatus200Json | AddPetStatus200Xml

/**
 * @type object
 */
export type AddPetStatus405 = {
  /**
   * @type integer | undefined
   */
  code?: number
  /**
   * @type string | undefined
   */
  message?: string
}

/**
 * @description Create a new pet in the store
 * @type object
 */
export type AddPetJsonData = AddPetRequest

/**
 * @description Create a new pet in the store
 * @type object
 */
export type AddPetXmlData = Pet

/**
 * @description Create a new pet in the store
 * @type object
 */
export type AddPetFormUrlEncodedData = Pet

export type AddPetData = AddPetJsonData | AddPetXmlData | AddPetFormUrlEncodedData

/**
 * @type object
 */
export type AddPetRequestConfig = {
  data?: AddPetData
  pathParams?: never
  queryParams?: never
  headerParams?: never
  /**
   * @type string
   */
  url: '/pet'
}

/**
 * @type object
 */
export type AddPetResponses = {
  '200': AddPetStatus200
  '405': AddPetStatus405
}

/**
 * @description Union of all possible responses
 */
export type AddPetResponse = AddPetStatus200 | AddPetStatus405