// examples/client/src/gen7/models/ts/petController/UploadFile.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ApiResponse } from '../ApiResponse.ts'

/**
 * @description ID of pet to update
 *
 * Format: `int64`
 * @type integer
 */
export type UploadFilePathPetId = bigint

/**
 * @description Additional Metadata
 * @type string | undefined
 */
export type UploadFileQueryAdditionalMetadata = string | undefined

/**
 * @type object
 */
export type UploadFileJsonData = {
  /**
   * @description URL of the image to upload
   *
   * Format: `uri`
   * @type string
   */
  url: string
}

/**
 * @type object
 */
export type UploadFileFormData = {
  /**
   * @description
   * Format: `binary`
   * @type string
   */
  file: Blob
}

export type UploadFileData = UploadFileJsonData | UploadFileFormData

/**
 * @type object
 */
export type UploadFileRequestConfig = {
  data?: UploadFileData
  /**
   * @type object
   */
  pathParams: {
    petId: UploadFilePathPetId
  }
  /**
   * @type object | undefined
   */
  queryParams?: {
    additionalMetadata?: UploadFileQueryAdditionalMetadata
  }
  headerParams?: never
  /**
   * @type string
   */
  url: `/pet/${string}/uploadImage`
}

/**
 * @type object
 */
export type UploadFileResponses = {
  /**
   * @type object
   */
  '200': ApiResponse
}

/**
 * @description Union of all possible responses
 */
export type UploadFileResponse = ApiResponse